Refactor ML data pipeline to use FastAPI, add Docker support, and restructure code for improved maintainability
