+++ ml_data_pipeline/pyproject.toml
@@ -15,15 +15,14 @@ omegaconf = "^2.3.0"
 pydantic = "^2.9.2"
 loguru = "^0.7.2"
 
-[tool.poetry.scripts]
-ml-data-pipeline = "ml_data_pipeline.main:main"
-
 [tool.poetry.group.dev.dependencies]
 pdoc = "^15.0.0"
 mypy = "^1.13.0"
 ruff = "^0.7.4"
 pytest = "^8.3.3"
 invoke = "^2.2.0"
+pydantic = "^2.9.2"
+fastapi = "^0.70.0"
 
 [build-system]
 requires = ["poetry-core"]
